# Reduce amount of c.Next calls in loader