Main bottleneck now is the amount of c.Next calls (on retain list of 1K items I see millions of .Next calls). 
- Reduced 100x amount of c.Next calls by removing next part: https://github.com/ledgerwatch/turbo-geth/compare/skip_too_much_ih?expand=1#diff-c75c86eeedd7661f1a471993b71b8774L419  - this logic is correct for `c` cursor, but not applicable for `ih` cursor. If `ih` cursor after jump to short prefix returns long `ihK` - it's still valid and can be used - no reason of jump to a next account. 
- Another minor improvement - reduce the amount of `.SeekTo` calls: funcion `keyIsBefore` returns true if keys are equal, but if keys are equal - no reason to call `.SeekTo`. Introduced function `keyIsBeforeOrEqual` and distinguish from `keyIsBefore`. 

Timings:
- Unwind of 1 block (300 items in RetainList) speedup from 5 sec to 200ms 
- Unwind of 1000 block (80K items in RetainList) speedup from 30 sec to 15 sec

- Bottleneck now moved to too-much small jumps over state after use of IH, will solve it in next PR.